requirements.txt
flask==2.3.3
flask-login==0.6.2
requests==2.31.0
werkzeug==2.3.7
supabase==2.0.3
python-dotenv==1.0.0
